updated max to spend label
